fix(events): isolate event form search params and bound public calendars
