+++ app/events/page.tsx
@@ -70,7 +70,13 @@ export default async function EventsPage() {
   // Fetch event calendars — for non-members, only include calendars with public events
   const { data: calendarsRaw, error: calendarsError } = isMember
     ? await supabase.from("event_calendars").select("*").order("name", { ascending: true })
-    : await supabase.from("event_calendars").select("*, events!inner(*)").eq("events.is_private", false).order("name", { ascending: true });
+    : await supabase
+        .from("event_calendars")
+        .select("*, events!inner(id)")
+        .eq("events.is_private", false)
+        .gte("events.start_time", oneYearAgo)
+        .lte("events.start_time", oneYearAhead)
+        .order("name", { ascending: true });
   if (calendarsError) {
     console.error("Failed to fetch event calendars:", calendarsError);
   }
